backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Caching
redis==5.0.1

# HTTP Clients
httpx==0.26.0
aiohttp==3.9.1

# AI Integration
google-generativeai==0.8.3

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS
python-jose[cryptography]==3.3.0

# Utilities
python-dateutil==2.8.2
